+++ packages/md/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @better-translate/md
 
+## 2.1.0
+
+### Minor Changes
+
+- Read me files fixed
+- Major changes for landing page, fixing readme files, versions, more stuff
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @better-translate/core@2.2.0
+
 ## 2.0.0
 
 ### Major Changes